Make Kafka the default storage implementation.
