+++ api-java/src/main/resources/config/kafka-producer.properties
